Update realtime-triggers.md
The "long" url with region in it for SQS will cause errors, as the Java SDK will not be able to connect with the long version.  Since you need to give a region anyhow, you do NOT want the long version of the queue URL. You should use the version of the url that is returned by * aws sqs list-queues --region XX-XXXX-XX", which does NOT have the region in the url.